update listen_handler to parse with atol instead of atoll
use of long long and atoll to parse a port number was deemed overkill.

+++ ldms/src/ldmsd/ldmsd_request.c
@@ -8799,7 +8799,7 @@ static int listen_handler(ldmsd_req_ctxt_t reqc)
 {
 	ldmsd_listen_t listen;
 	char *xprt, *port, *host, *auth, *attr_name, *quota, *rx_limit;
-	long long port_no = -1;
+	long port_no = -1;
 	xprt = port = host = auth = quota = rx_limit = NULL;
 
 	attr_name = "xprt";
@@ -8808,7 +8808,7 @@ static int listen_handler(ldmsd_req_ctxt_t reqc)
 		goto einval;
 	port = ldmsd_req_attr_str_value_get_by_id(reqc, LDMSD_ATTR_PORT);
 	if (port) {
-		port_no = atoll(port);
+		port_no = atol(port);
 		if (port_no < 1 || port_no > USHRT_MAX) {
 			reqc->errcode = EINVAL;
 			(void) snprintf(reqc->line_buf, reqc->line_len,
